+++ packages/extension-base/src/utils/account/transform.ts
@@ -98,6 +98,8 @@ export const getAccountSignMode = (address: string, _meta?: KeyringPair$Meta): A
           }
         } else if (meta.isReadOnly) {
           return AccountSignMode.READ_ONLY;
+        } else if (meta.isMultisig) {
+          return AccountSignMode.MULTISIG;
         } else {
           return AccountSignMode.QR;
         }
@@ -246,6 +248,19 @@ const OPEN_GOV_ACTIONS: ExtrinsicType[] = [
   ExtrinsicType.GOV_UNLOCK_VOTE
 ];
 
+const MULTISIG_ACTIONS: ExtrinsicType[] = [
+  ExtrinsicType.MULTISIG_APPROVE_TX,
+  ExtrinsicType.MULTISIG_EXECUTE_TX,
+  ExtrinsicType.MULTISIG_CANCEL_TX,
+  ExtrinsicType.MULTISIG_INIT_TX
+];
+
+const SUBSTRATE_PROXY_ACTION: ExtrinsicType[] = [
+  ExtrinsicType.ADD_SUBSTRATE_PROXY_ACCOUNT,
+  ExtrinsicType.REMOVE_SUBSTRATE_PROXY_ACCOUNT,
+  ExtrinsicType.SUBSTRATE_PROXY_INIT_TX
+];
+
 const OTHER_ACTIONS: ExtrinsicType[] = [
   ExtrinsicType.TRANSFER_XCM,
   ExtrinsicType.SEND_NFT,
@@ -268,6 +283,8 @@ export const getAccountTransactionActions = (signMode: AccountSignMode, networkT
           ...EARN_VMANTA_ACTIONS,
           ...CLAIM_AVAIL_BRIDGE,
           ...OPEN_GOV_ACTIONS,
+          ...MULTISIG_ACTIONS,
+          ...SUBSTRATE_PROXY_ACTION,
           ...OTHER_ACTIONS
         ];
       case AccountChainType.ETHEREUM:
@@ -278,6 +295,7 @@ export const getAccountTransactionActions = (signMode: AccountSignMode, networkT
           ...EARN_STDOT_ACTIONS,
           ...OTHER_ACTIONS,
           ...CLAIM_AVAIL_BRIDGE,
+          ...SUBSTRATE_PROXY_ACTION,
           ...EVM_ACTIONS
         ];
       case AccountChainType.TON:
@@ -308,6 +326,8 @@ export const getAccountTransactionActions = (signMode: AccountSignMode, networkT
           ...EARN_VMANTA_ACTIONS,
           ...CLAIM_AVAIL_BRIDGE,
           ...OPEN_GOV_ACTIONS,
+          ...MULTISIG_ACTIONS,
+          ...SUBSTRATE_PROXY_ACTION,
           ...OTHER_ACTIONS
         ];
       case AccountChainType.ETHEREUM:
@@ -333,6 +353,31 @@ export const getAccountTransactionActions = (signMode: AccountSignMode, networkT
       case AccountChainType.BITCOIN:
         return [];
     }
+  } else if (signMode === AccountSignMode.MULTISIG) {
+    switch (networkType) {
+      case AccountChainType.SUBSTRATE:
+        return [
+          ...BASE_TRANSFER_ACTIONS,
+          ...NATIVE_STAKE_ACTIONS,
+          ...POOL_STAKE_ACTIONS,
+          // ...EARN_VDOT_ACTIONS, // todo: consider support this
+          // ...EARN_LDOT_ACTIONS,
+          // ...EARN_SDOT_ACTIONS,
+          // ...EARN_QDOT_ACTIONS,
+          // ...EARN_VMANTA_ACTIONS,
+          ...OPEN_GOV_ACTIONS,
+          ...SUBSTRATE_PROXY_ACTION,
+          ExtrinsicType.SEND_NFT
+        ];
+      case AccountChainType.ETHEREUM:
+        return [];
+      case AccountChainType.TON:
+        return [];
+      case AccountChainType.CARDANO:
+        return [];
+      case AccountChainType.BITCOIN:
+        return [];
+    }
   } else if (signMode === AccountSignMode.GENERIC_LEDGER) {
     switch (networkType) {
       case AccountChainType.SUBSTRATE:
@@ -346,6 +391,8 @@ export const getAccountTransactionActions = (signMode: AccountSignMode, networkT
           ...EARN_SDOT_ACTIONS,
           // ...EARN_QDOT_ACTIONS,
           ...OPEN_GOV_ACTIONS,
+          ...MULTISIG_ACTIONS,
+          ...SUBSTRATE_PROXY_ACTION,
           ...OTHER_ACTIONS
         ];
       case AccountChainType.ETHEREUM:
@@ -542,6 +589,8 @@ export const convertAccountProxyType = (accountSignMode: AccountSignMode): Accou
       return AccountProxyType.LEDGER;
     case AccountSignMode.QR:
       return AccountProxyType.QR;
+    case AccountSignMode.MULTISIG:
+      return AccountProxyType.MULTISIG;
     case AccountSignMode.READ_ONLY:
       return AccountProxyType.READ_ONLY;
     case AccountSignMode.INJECTED: