+++ packages/extension-koni-ui/src/Popup/Confirmations/variants/Transaction/variants/LeavePool.tsx
@@ -38,12 +38,12 @@ const Component: React.FC<Props> = (props: Props) => {
           value={data.amount}
         />
 
-        <MetaInfo.Number
+        {!transaction.wrappingStatus && <MetaInfo.Number
           decimals={decimals}
-          label={t('ui.TRANSACTION.Confirmations.LeavePool.estimatedFee')}
+          label={t('ui.TRANSACTION.Confirmations.LeavePool.networkFee')}
           suffix={symbol}
           value={transaction.estimateFee?.value || 0}
-        />
+        />}
       </MetaInfo>
     </div>
   );