+++ packages/extension-koni-ui/src/Popup/Transaction/variants/Governance/ReferendumVote/ReferendumAbstainVote.tsx
@@ -61,7 +61,7 @@ const Component = (props: ComponentProps): React.ReactElement<ComponentProps> =>
   const abstainRenderKey = 'abstain_vote_amount';
   const [abstainAmountRenderKey, setAbstainAmountRenderKey] = useState<string>(abstainRenderKey);
 
-  const onPreCheck = usePreCheckAction(fromValue);
+  const onPreCheck = usePreCheckAction({ chain: chainValue, address: fromValue });
   const { onError, onSuccess } = useHandleSubmitTransaction();
 
   const { voteMap = new Map<string, ReferendumVoteDetail>(),
@@ -92,9 +92,7 @@ const Component = (props: ComponentProps): React.ReactElement<ComponentProps> =>
     };
 
     handleVote(voteRequest)
-      .then((tx) => {
-        onSuccess(tx);
-      })
+      .then(onSuccess)
       .catch(onError)
       .finally(() => setLoading(false));
   }, [chainValue, defaultData.referendumId, defaultData.track, onError, onSuccess]);