+++ packages/extension-koni-ui/src/utils/transaction/detectType.ts
@@ -50,3 +50,15 @@ export const isTypeGov = (txType: ExtrinsicType) => [
   ExtrinsicType.GOV_UNVOTE,
   ExtrinsicType.GOV_UNLOCK_VOTE
 ].includes(txType);
+
+export const isTypeManageSubstrateProxy = (txType: ExtrinsicType) => [
+  ExtrinsicType.ADD_SUBSTRATE_PROXY_ACCOUNT,
+  ExtrinsicType.REMOVE_SUBSTRATE_PROXY_ACCOUNT
+].includes(txType);
+
+export const isTypeMultisig = (txType: ExtrinsicType) => [
+  ExtrinsicType.MULTISIG_APPROVE_TX,
+  ExtrinsicType.MULTISIG_EXECUTE_TX,
+  ExtrinsicType.MULTISIG_CANCEL_TX,
+  ExtrinsicType.MULTISIG_INIT_TX
+].includes(txType);